SignUp module

+++ app/src/main/java/com/app/okra/models/InsightResponse.kt
@@ -1,5 +1,5 @@
 package com.app.okra.models
 
 class InsightResponse(
-        var insightData: ArrayList<Int>?= null
+        var insightData: ArrayList<Float>?= null
 )
